Ensuring check the actual new pod got migrated
In order to test the pod migration when the current node is cordoned, we
get the list of running pods and cordon the node for the first pod in
the list.

Once the pod is migrated, we get the new list and verify the node for
the first pod has changed.

The problem is the first pod in the new list is not necessarily the new
pod and, even more, it may be more than one pod were running on the
cordoned node which may lead us to think the node was not migrated.

With this change, we ensure we check the node for the new (migrated) pod
in the replica set.